[SD-WebUI/Forge] SEED LOGIC CHANGE: Execute fix_seed before calculating TIPO seed. Only affects SD-WebUI/Forge.
Reason: This will fix image re-creation in SD-WebUI/Forge when restoring prompts. Images generated prior to this change cannot be reproduced when TIPO was used with the `Upsampling timing` = `BEFORE` setting.

Prior to this change it was possible to reproduce an image by resetting the txt2img seed back to -1 after prompt restoration (if it was -1 during the initial generation).

With this change resetting to -1 is not needed anymore. The TIPO seed logic has been fixed to use the correct txt2img seed (but this will make images generated prior to this change not reproducible due to logic changes).